# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "DayScheduleRead": {
        "description": "No `date` field - the date is already the key this value sits under in\nScheduleResponse.days, so it isn't duplicated here.",
        "properties": {
          "cancelled": {
            "items": {
              "$ref": "#/components/schemas/EntryRefRead"
            },
            "title": "Cancelled",
            "type": "array"
          },
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ResolvedEntryRead"
            },
            "title": "Entries",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus"
          }
        },
        "required": [
          "status",
          "entries",
          "cancelled",
          "completed"
        ],
        "title": "DayScheduleRead",
        "type": "object"
      },
      "DayStatus": {
        "description": "Day-level status is a presentation convenience for compact rendering\n(e.g. a single calendar-cell badge) - `entries` and `cancelled` are the\nauthoritative data. Clients must not branch on this status alone in a\ndetailed view: a SCHEDULED day can still carry unrelated cancelled\nentries (see resolve()'s mixed-day precedence), and SUBSTITUTED only\nmeans *at least one* surviving entry that day is a substitution.\n\nThese string values are part of the public wire contract - app/schemas/\nimports this enum directly for the API response, so renaming a member\nhere changes the API, not just an internal detail.",
        "enum": [
          "empty",
          "scheduled",
          "cancelled",
          "substituted"
        ],
        "title": "DayStatus",
        "type": "string"
      },
...
      "EntryRefRead": {
        "description": "Minimal reference to another entry - used both for what a substitution\nreplaced and what a cancellation cancelled, since both need exactly the\nsame information: which entry, and what to call it. Never None.entry_id;\nthe whole reference is Optional at the call site instead, for the cases\nwhere there is nothing to reference at all.",
        "properties": {
          "entry_id": {
            "format": "uuid",
            "title": "Entry Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "entry_id",
          "name"
        ],
        "title": "EntryRefRead",
        "type": "object"
      },
      "EntryStatus": {
        "enum": [
          "scheduled",
          "substituted"
        ],
        "title": "EntryStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolvedEntryRead": {
        "properties": {
          "completion_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Id"
          },
          "entry_id": {
            "format": "uuid",
            "title": "Entry Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "replaced": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntryRefRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "workout_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Id"
          }
        },
        "required": [
          "entry_id",
          "workout_id",
          "name",
          "notes",
          "status",
          "replaced",
          "completion_id"
        ],
        "title": "ResolvedEntryRead",
        "type": "object"
      },
      "ScheduleEntryCreate": {
        "additionalProperties": false,
        "properties": {
          "day_of_week": {
            "anyOf": [
              {
                "maximum": 7.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week"
          },
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "name_override": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Override"
          },
          "on_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Date"
          },
          "replaces_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Entry Id"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "workout_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Id"
          }
        },
        "title": "ScheduleEntryCreate",
        "type": "object"
      },
      "ScheduleEntryRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "day_of_week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week"
          },
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Override"
          },
          "on_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Date"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "replaces_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Entry Id"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "workout_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Id"
          }
        },
        "required": [
          "id",
          "plan_id",
          "workout_id",
          "day_of_week",
          "on_date",
          "replaces_entry_id",
          "name_override",
          "starts_on",
          "ends_on",
          "created_at"
        ],
        "title": "ScheduleEntryRead",
        "type": "object"
      },
      "ScheduleEntryUpdate": {
        "additionalProperties": false,
        "properties": {
          "day_of_week": {
            "anyOf": [
              {
                "maximum": 7.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week"
          },
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "name_override": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Override"
          },
          "on_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Date"
          },
          "replaces_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Entry Id"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          },
          "workout_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Id"
          }
        },
        "title": "ScheduleEntryUpdate",
        "type": "object"
      },
      "ScheduleResponse": {
        "properties": {
          "days": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DayScheduleRead"
            },
            "title": "Days",
            "type": "object"
          }
        },
        "required": [
          "days"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "follow-through API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/plans/{plan_id}/schedule": {
      "get": {
        "operationId": "get_schedule_plans__plan_id__schedule_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Schedule",
        "tags": [
          "schedule"
        ]
      }
    },
    "/plans/{plan_id}/schedule-entries": {
      "get": {
        "operationId": "list_entries_plans__plan_id__schedule_entries_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleEntryRead"
                  },
                  "title": "Response List Entries Plans  Plan Id  Schedule Entries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Entries",
        "tags": [
          "schedule-entries"
        ]
      },
      "post": {
        "operationId": "create_entry_plans__plan_id__schedule_entries_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleEntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Entry",
        "tags": [
          "schedule-entries"
        ]
      }
    },
    "/plans/{plan_id}/schedule-entries/{entry_id}": {
      "delete": {
        "operationId": "delete_entry_plans__plan_id__schedule_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Entry",
        "tags": [
          "schedule-entries"
        ]
      },
      "patch": {
        "operationId": "update_entry_plans__plan_id__schedule_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleEntryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleEntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Entry",
        "tags": [
          "schedule-entries"
        ]
      }
    },
...
  }
}